Disable our custom crash handler three months after a release. This implements issue #414.
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@11662 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/crashrpt/excprpt.cpp
@@ -234,7 +234,7 @@ string CExceptionReport::getSymbolFile(int index)
 {
    string ret;
 
-   if (0 < index && index < m_symFiles.size())
+   if (0 < index && index < (int)m_symFiles.size())
       ret = m_symFiles[index];
 
    return ret;
